+++ backend/src/modules/tickets/tickets.module.ts
@@ -6,6 +6,7 @@ import { TicketCategoryConfig } from './entities/ticket-category.entity';
 import { TicketKeywordRule } from './entities/ticket-keyword-rule.entity';
 import { TechAttendance } from './entities/tech-attendance.entity';
 import { OfficeDay } from './entities/office-day.entity';
+import { TicketEvent } from './entities/ticket-event.entity';
 import { TicketService } from './services/ticket.service';
 import { TicketSettingsService } from './services/ticket-settings.service';
 import { AttendanceService } from './services/attendance.service';
@@ -25,6 +26,7 @@ import { RoleDefinitionEntity } from '../users/entities/role-definition.entity';
       TicketKeywordRule,
       TechAttendance,
       OfficeDay,
+      TicketEvent,
       User,
       RoleDefinitionEntity,
     ]),